requirements.txt
numpy
matplotlib
pillow
torch
torchvision
scikit-learn

